Move functions into schema files
