+++ apps/ink_api/lib/ink_api/graphql/schema/types/user.ex
@@ -12,14 +12,6 @@ defmodule InkApi.Schema.Types.User do
   end
 
   object :user_mutations do
-    field :log_in_with_provider, type: :session do
-      arg :code, non_null(:string)
-      arg :provider, non_null(:string)
-      arg :redirect_uri, non_null(:string)
-
-      resolve &Resolver.User.log_in_with_provider/2
-    end
-
     field :login, type: :session do
       arg :email, non_null(:string)
       arg :password, non_null(:string)
